# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Optional map of Modbus register addresses/names to values.
 */
export type ModbusReadingIngestBodyRegisters = {
  [key: string]: string | number | boolean | null;
};

/**
 * Optional decoded values sent by the reader.
 */
export type ModbusReadingIngestBodyValues = { [key: string]: unknown };

export interface ModbusReadingIngestBody {
  /** Optional reader/device identifier. Defaults to "trb246" when omitted. */
  deviceId?: string;
  /** Optional timestamp generated by the reader. */
  timestamp?: string;
  /** Optional map of Modbus register addresses/names to values. */
  registers?: ModbusReadingIngestBodyRegisters;
  /** Optional decoded values sent by the reader. */
  values?: ModbusReadingIngestBodyValues;
  /** Optional nested raw payload from the reader. */
  payload?: unknown;
  [key: string]: unknown;
}

export type ModbusReadingRawPayload = { [key: string]: unknown };

export type ModbusDecodedValuesStatus =
  (typeof ModbusDecodedValuesStatus)[keyof typeof ModbusDecodedValuesStatus];

export const ModbusDecodedValuesStatus = {
  decoded: "decoded",
  contains_unknown_registers: "contains_unknown_registers",
  contains_invalid_registers: "contains_invalid_registers",
  no_registers: "no_registers",
} as const;

export type ModbusDecodedRegisterStatus =
  (typeof ModbusDecodedRegisterStatus)[keyof typeof ModbusDecodedRegisterStatus];

export const ModbusDecodedRegisterStatus = {
  decoded: "decoded",
  unknown: "unknown",
  invalid: "invalid",
} as const;

export interface ModbusDecodedRegister {
  address: string;
  name: string;
  /** @nullable */
  unit: string | null;
  status: ModbusDecodedRegisterStatus;
  /** @nullable */
  value: string | number | boolean | null;
  rawValue: unknown;
  displayValue?: string;
  error?: string;
}

export type ModbusDecodedValuesProvidedValues = { [key: string]: unknown };

export interface ModbusDecodedValues {
  status: ModbusDecodedValuesStatus;
  registers: ModbusDecodedRegister[];
  providedValues: ModbusDecodedValuesProvidedValues;
}

export interface ModbusReading {
  id: number;
  deviceId: string;
  /** @nullable */
  source: string | null;
  parsingStatus: string;
  rawPayload: ModbusReadingRawPayload;
  decodedValues: ModbusDecodedValues;
  receivedAt: string;
}

export interface ModbusReadingAck {
  accepted: boolean;
  reading: ModbusReading;
}

export interface ModbusReadingsList {
  readings: ModbusReading[];
}

export interface ErrorResponse {
  error: string;
}

export type NotificationChannelConfigInApp = {
  enabled: boolean;
};

export type NotificationChannelConfigWebhook = {
  enabled: boolean;
  url: string;
};

export type NotificationChannelConfigEmail = {
  enabled: boolean;
  to: string;
};

export interface NotificationChannelConfig {
  inApp: NotificationChannelConfigInApp;
  webhook: NotificationChannelConfigWebhook;
  email: NotificationChannelConfigEmail;
}

export interface NotificationPreferences {
  id: number;
  scope: string;
  enabled: boolean;
  /** @minimum 1 */
  thresholdMinutes: number;
  /** @minimum 1 */
  cooldownMinutes: number;
  channels: NotificationChannelConfig;
  updatedAt: string;
}

export interface NotificationPreferencesResponse {
  preferences: NotificationPreferences;
}

export interface UpdateNotificationPreferencesBody {
  enabled: boolean;
  /**
   * @minimum 1
   * @maximum 1440
   */
  thresholdMinutes: number;
  /**
   * @minimum 1
   * @maximum 1440
   */
  cooldownMinutes: number;
  channels: NotificationChannelConfig;
}

export type DeviceAlertChannelDispatchChannel =
  (typeof DeviceAlertChannelDispatchChannel)[keyof typeof DeviceAlertChannelDispatchChannel];

export const DeviceAlertChannelDispatchChannel = {
  inApp: "inApp",
  webhook: "webhook",
  email: "email",
} as const;

export type DeviceAlertChannelDispatchStatus =
  (typeof DeviceAlertChannelDispatchStatus)[keyof typeof DeviceAlertChannelDispatchStatus];

export const DeviceAlertChannelDispatchStatus = {
  delivered: "delivered",
  skipped: "skipped",
  failed: "failed",
} as const;

export interface DeviceAlertChannelDispatch {
  channel: DeviceAlertChannelDispatchChannel;
  status: DeviceAlertChannelDispatchStatus;
  detail?: string;
}

export type DeviceAlertEventSeverity =
  (typeof DeviceAlertEventSeverity)[keyof typeof DeviceAlertEventSeverity];

export const DeviceAlertEventSeverity = {
  warning: "warning",
  fault: "fault",
  resolved: "resolved",
} as const;

export interface DeviceAlertEvent {
  id: number;
  deviceId: string;
  severity: DeviceAlertEventSeverity;
  minutesSinceData: number;
  thresholdMinutes: number;
  message: string;
  dispatch: DeviceAlertChannelDispatch[];
  /** @nullable */
  acknowledgedAt: string | null;
  createdAt: string;
}

export interface DeviceAlertEventsList {
  events: DeviceAlertEvent[];
}

export interface EvaluateAlertsResponse {
  evaluated: number;
  dispatched: number;
}

export interface AlertTestBody {
  deviceId?: string;
}

export interface AlertTestResponse {
  event: DeviceAlertEvent;
}

export type ListModbusReadingsParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * Optional device identifier to filter readings by a single TRB246 gateway.
   * @minLength 1
   */
  deviceId?: string;
  /**
   * Optional ISO-8601 timestamp; only readings received at or after this time are returned.
   */
  since?: string;
  /**
   * Optional ISO-8601 timestamp; only readings received at or before this time are returned.
   */
  until?: string;
};

export type ListAlertEventsParams = {
  /**
   * @minimum 1
   * @maximum 200
   */
  limit?: number;
  /**
   * Optional ISO-8601 timestamp; only events created at or after this time are returned.
   */
  since?: string;
};