# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: modbus
    description: Modbus reader ingestion operations
  - name: alerts
    description: Stale-device alert preferences and event log
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /modbus/readings:
    get:
      operationId: listModbusReadings
      tags: [modbus]
      summary: List recent Modbus readings
      description: Returns recent payloads received from a Modbus reader.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: deviceId
          in: query
          required: false
          description: Optional device identifier to filter readings by a single TRB246 gateway.
          schema:
            type: string
            minLength: 1
        - name: since
          in: query
          required: false
          description: Optional ISO-8601 timestamp; only readings received at or after this time are returned.
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          required: false
          description: Optional ISO-8601 timestamp; only readings received at or before this time are returned.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Recent readings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModbusReadingsList"
    post:
      operationId: createModbusReading
      tags: [modbus]
      summary: Receive Modbus reader data
      description: |
        Accepts HTTP JSON payloads from a Teltonika TRB246 or other Modbus reader. Requests must include the shared device token in the x-device-key header or as a Bearer token.

        During a token rotation, the server temporarily accepts both the current token (`MODBUS_INGEST_TOKEN`) and any tokens listed in `MODBUS_INGEST_TOKEN_PREVIOUS` (comma-separated). Requests authenticated with a previous token are logged so operators can confirm migration before retiring the old token.
      security:
        - DeviceKeyAuth: []
        - DeviceBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModbusReadingIngestBody"
      responses:
        "200":
          description: Reading accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModbusReadingAck"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid device token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Device ingest token is not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /alerts/preferences:
    get:
      operationId: getAlertPreferences
      tags: [alerts]
      summary: Get notification preferences
      responses:
        "200":
          description: Current preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
    put:
      operationId: updateAlertPreferences
      tags: [alerts]
      summary: Update notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferencesBody"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferencesResponse"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /alerts/events:
    get:
      operationId: listAlertEvents
      tags: [alerts]
      summary: List recent alert events
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: since
          in: query
          required: false
          description: Optional ISO-8601 timestamp; only events created at or after this time are returned.
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Recent alert events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceAlertEventsList"
  /alerts/test:
    post:
      operationId: sendTestAlert
      tags: [alerts]
      summary: Dispatch a test alert through configured channels
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertTestBody"
      responses:
        "200":
          description: Test alert dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertTestResponse"
  /alerts/evaluate:
    post:
      operationId: evaluateAlertsNow
      tags: [alerts]
      summary: Manually trigger the staleness evaluator
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateAlertsResponse"
components:
  securitySchemes:
    DeviceKeyAuth:
      type: apiKey
      in: header
      name: x-device-key
      description: Shared device token stored in the API server environment as MODBUS_INGEST_TOKEN. During rotation, comma-separated tokens in MODBUS_INGEST_TOKEN_PREVIOUS are also accepted.
    DeviceBearerAuth:
      type: http
      scheme: bearer
      description: Alternative bearer-token format using the same shared device token (current or previous during rotation).
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ModbusReadingIngestBody:
      type: object
      additionalProperties: true
      properties:
        deviceId:
          type: string
          description: Optional reader/device identifier. Defaults to "trb246" when omitted.
        timestamp:
          type: string
          format: date-time
          description: Optional timestamp generated by the reader.
        registers:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: integer
              - type: boolean
              - type: "null"
          description: Optional map of Modbus register addresses/names to values.
        values:
          type: object
          additionalProperties: true
          description: Optional decoded values sent by the reader.
        payload:
          description: Optional nested raw payload from the reader.
      minProperties: 1
    ModbusReading:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: string
        source:
          type: ["string", "null"]
        parsingStatus:
          type: string
        rawPayload:
          type: object
          additionalProperties: true
        decodedValues:
          $ref: "#/components/schemas/ModbusDecodedValues"
        receivedAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - source
        - parsingStatus
        - rawPayload
        - decodedValues
        - receivedAt
    ModbusDecodedValues:
      type: object
      properties:
        status:
          type: string
          enum:
            - decoded
            - contains_unknown_registers
            - contains_invalid_registers
            - no_registers
        registers:
          type: array
          items:
            $ref: "#/components/schemas/ModbusDecodedRegister"
        providedValues:
          type: object
          additionalProperties: true
      required:
        - status
        - registers
        - providedValues
    ModbusDecodedRegister:
      type: object
      properties:
        address:
          type: string
        name:
          type: string
        unit:
          type: ["string", "null"]
        status:
          type: string
          enum:
            - decoded
            - unknown
            - invalid
        value:
          type:
            - string
            - number
            - boolean
            - "null"
        rawValue: {}
        displayValue:
          type: string
        error:
          type: string
      required:
        - address
        - name
        - unit
        - status
        - value
        - rawValue
    ModbusReadingAck:
      type: object
      properties:
        accepted:
          type: boolean
        reading:
          $ref: "#/components/schemas/ModbusReading"
      required:
        - accepted
        - reading
    ModbusReadingsList:
      type: object
      properties:
        readings:
          type: array
          items:
            $ref: "#/components/schemas/ModbusReading"
      required:
        - readings
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    NotificationChannelConfig:
      type: object
      properties:
        inApp:
          type: object
          properties:
            enabled:
              type: boolean
          required: [enabled]
        webhook:
          type: object
          properties:
            enabled:
              type: boolean
            url:
              type: string
          required: [enabled, url]
        email:
          type: object
          properties:
            enabled:
              type: boolean
            to:
              type: string
          required: [enabled, to]
      required: [inApp, webhook, email]
    NotificationPreferences:
      type: object
      properties:
        id:
          type: integer
        scope:
          type: string
        enabled:
          type: boolean
        thresholdMinutes:
          type: integer
          minimum: 1
        cooldownMinutes:
          type: integer
          minimum: 1
        channels:
          $ref: "#/components/schemas/NotificationChannelConfig"
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - scope
        - enabled
        - thresholdMinutes
        - cooldownMinutes
        - channels
        - updatedAt
    NotificationPreferencesResponse:
      type: object
      properties:
        preferences:
          $ref: "#/components/schemas/NotificationPreferences"
      required: [preferences]
    UpdateNotificationPreferencesBody:
      type: object
      properties:
        enabled:
          type: boolean
        thresholdMinutes:
          type: integer
          minimum: 1
          maximum: 1440
        cooldownMinutes:
          type: integer
          minimum: 1
          maximum: 1440
        channels:
          $ref: "#/components/schemas/NotificationChannelConfig"
      required: [enabled, thresholdMinutes, cooldownMinutes, channels]
    DeviceAlertChannelDispatch:
      type: object
      properties:
        channel:
          type: string
          enum: [inApp, webhook, email]
        status:
          type: string
          enum: [delivered, skipped, failed]
        detail:
          type: string
      required: [channel, status]
    DeviceAlertEvent:
      type: object
      properties:
        id:
          type: integer
        deviceId:
          type: string
        severity:
          type: string
          enum: [warning, fault, resolved]
        minutesSinceData:
          type: integer
        thresholdMinutes:
          type: integer
        message:
          type: string
        dispatch:
          type: array
          items:
            $ref: "#/components/schemas/DeviceAlertChannelDispatch"
        acknowledgedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - severity
        - minutesSinceData
        - thresholdMinutes
        - message
        - dispatch
        - acknowledgedAt
        - createdAt
    DeviceAlertEventsList:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/DeviceAlertEvent"
      required: [events]
    EvaluateAlertsResponse:
      type: object
      properties:
        evaluated:
          type: integer
        dispatched:
          type: integer
      required: [evaluated, dispatched]
    AlertTestBody:
      type: object
      properties:
        deviceId:
          type: string
    AlertTestResponse:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/DeviceAlertEvent"
      required: [event]