# lib/api-zod/src/generated/types/deviceAlertEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceAlertChannelDispatch } from "./deviceAlertChannelDispatch";
import type { DeviceAlertEventSeverity } from "./deviceAlertEventSeverity";

export interface DeviceAlertEvent {
  id: number;
  deviceId: string;
  severity: DeviceAlertEventSeverity;
  minutesSinceData: number;
  thresholdMinutes: number;
  message: string;
  dispatch: DeviceAlertChannelDispatch[];
  /** @nullable */
  acknowledgedAt: Date | null;
  createdAt: Date;
}