# lib/api-zod/src/generated/types/notificationChannelConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationChannelConfigEmail } from "./notificationChannelConfigEmail";
import type { NotificationChannelConfigInApp } from "./notificationChannelConfigInApp";
import type { NotificationChannelConfigWebhook } from "./notificationChannelConfigWebhook";

export interface NotificationChannelConfig {
  inApp: NotificationChannelConfigInApp;
  webhook: NotificationChannelConfigWebhook;
  email: NotificationChannelConfigEmail;
}