# lib/api-zod/src/generated/types/notificationPreferences.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationChannelConfig } from "./notificationChannelConfig";

export interface NotificationPreferences {
  id: number;
  scope: string;
  enabled: boolean;
  /** @minimum 1 */
  thresholdMinutes: number;
  /** @minimum 1 */
  cooldownMinutes: number;
  channels: NotificationChannelConfig;
  updatedAt: Date;
}