# lib/api-zod/src/generated/types/notificationPreferencesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationPreferences } from "./notificationPreferences";

export interface NotificationPreferencesResponse {
  preferences: NotificationPreferences;
}