# lib/api-zod/src/generated/types/updateNotificationPreferencesBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationChannelConfig } from "./notificationChannelConfig";

export interface UpdateNotificationPreferencesBody {
  enabled: boolean;
  /**
   * @minimum 1
   * @maximum 1440
   */
  thresholdMinutes: number;
  /**
   * @minimum 1
   * @maximum 1440
   */
  cooldownMinutes: number;
  channels: NotificationChannelConfig;
}